some progress

+++ task_manager/src/main/java/com/acg/task_manager/repository/TaskRepository.java
@@ -3,6 +3,6 @@ package com.acg.task_manager.repository;
 import org.springframework.data.repository.CrudRepository;
 import com.acg.task_manager.model.*;
 
-public interface TaskRepository extends CrudRepository<Task, Integer> {
+public interface TaskRepository extends CrudRepository<Task, Long> {
 
 }
